feat: implement escrow contract logic with event schemas and comprehensive documentation
